Change gravity of customer TextView to left in counter fragment
The gravity of the customer TextView in the counter fragment was changed from center to left. This was done to improve the readability of the customer's name, which can be long and wrap to multiple lines.

In addition, the app name was updated from "نقطةبيع" to "نقطة بيع " in the strings.xml file. The gravity of the total sales TextView in the today fragment was also changed to left. Finally, the background tint of the export CSV and share report buttons in the reports enhanced fragment was changed to the accent color.